Add first pass parsing and data generation
We now are auto-generating a file that contains all the data as accessor
functions with the correct names for the enzymes (in elixir compatible
casing).

This was just a first pass though, so I do want to change how the data
is structured. For example, removing duplicated data, as well as
changing it so that e.g. pattern matching on a group of enzymes is
easier.