Finalize a beta version of the data layout
I played around a bit to find out what would work the best in terms of
the data structure to use. I landed on structs early on, but I wasn't
sure how they were best used. For example, should I create a struct for
_each_ of the enzymes?

I landed on keeping the struct simple for the sake of type information
in functions that require an enzyme. That way we can use something like:
``` elixir
def digest(dna, %Bio.Restriction.Enzyme{pattern: split}) do ....
```

As a contrived example, of course.