Restructure for polymer concept as top level
The idea of the polymers is that each sub-module represents a
distinctive type which can then house the relative struct for that
polymer.

Looking at the implementation for DNA should make this clear. But in
general, I think that it makes sense to have the struct point to a
specific `polymer` map that can then be used to define the relevant
information for a more nuanced piece of data.

Wherever possible (which should be most places), there will be a
function for building the complex data from a simpler type. As with the
`Bio.Polymer.Dna.from_binary/2` function.

Additionally, anything in the module that can be defined to act on the
naive data type (binary) should be able to do so, in addition to
overloading for the case of the more complicated struct.

The goal is ultimatley to allow users the greatest degree of freedom in
deciding how they use the system. We won't force them to convert things
that don't need to be converted unless the operation actually requires
it.